moved code to wq.c; wait_enqueue assumes not enqueued
